Add modal to onClick
